fix: 型定義を１つのファイルに統合+名前付きエクスポートに修正
